never gonna give you up

+++ src/graphics.js
@@ -3,7 +3,7 @@ import Sk from "./skulpt.js";
 import { makeFunc, __name__, self } from "./utils.js";
 import PApplet from "./applet.js";
 
-const { int } = Sk.builtin;
+const { int_ } = Sk.builtin;
 const { buildClass } = Sk.misceval;
 
 function graphicsInit(self, width, height, applet) {
@@ -13,7 +13,7 @@ function graphicsInit(self, width, height, applet) {
 function graphicsClass($gbl, $loc) {
     $loc.__init__ = makeFunc(graphicsInit, [
         self,
-        { "width": int },
+        { "width": int_ },
         { "width": int },
         { "width": PApplet }
     ]);
